refactor: extract TangentData from JsonXModelLoader
